Revert "Reverted struct BagPocket capacity variable changes"
This reverts commit 1199f9fd3a80a883d066952eb3726d763f012718.

+++ include/item.h
@@ -26,7 +26,7 @@ struct Item
 struct BagPocket
 {
     struct ItemSlot *itemSlots;
-    u8 capacity;
+    u16 capacity;
 };
 
 extern const struct Item gItems[];
